+++ CHANGELOG.rst
@@ -8,6 +8,10 @@ manuelraa.yaml_folder_inventory Release Notes
 v1.0.1
 ======
 
+Breaking Changes
+----------------
+- Rename inventory plugin from :code:`manuelraa.inventory.yaml_folder` to :code:`manuelraa.yaml_folder.yaml_folder`
+
 Minor Changes
 -------------
 - Add type hints and basic docstrings
@@ -15,6 +19,7 @@ Minor Changes
 Bugfixes
 --------
 - Removed debug print from plugin
+- Fix empty host vars in :code:`main.yml` causes error because parsed as None
 
 
 v1.0.0